로그인 프로세스 및 관리자 (사용자 관리) 처리 [진행중]

+++ src/main/java/bb/toy/api/dto/member/ResultMemberDto.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class ResultMemberDto<T> {
-    private int count;
+    private Long count;
     private T data;
 }
